Move device locking into aspSUB20.
